+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- **`/features` endpoints**: both json and geojson responses.
+
 ## [1.6.0] - 2026-06-22
 
 ### Added